# API_DOCUMENTATION.json
{
    "openapi":  "3.0.1",
    "info":  {
                 "title":  "Finance Data Processing API",
                 "description":  "Backend API for Finance Dashboard with Role-Based Access Control",
                 "contact":  {
                                 "name":  "Md Shahbaz Imam",
                                 "email":  "mdshahbazimam02@gmail.com"
                             },
                 "version":  "1.0.0"
             },
    "servers":  [
                    {
                        "url":  "http://localhost:8080",
                        "description":  "Default Server"
                    }
                ],
    "security":  [
                     {
                         "bearerAuth":  [

                                        ]
                     }
                 ],
...
    "paths":  {
...
                  "/api/users/{id}/status":  {
                                                 "patch":  {
                                                               "tags":  [
                                                                            "User Management"
                                                                        ],
                                                               "summary":  "Update user status (active/inactive)",
                                                               "operationId":  "updateStatus",
                                                               "parameters":  [
                                                                                  {
                                                                                      "name":  "id",
                                                                                      "in":  "path",
                                                                                      "required":  true,
                                                                                      "schema":  {
                                                                                                     "type":  "integer",
                                                                                                     "format":  "int64"
                                                                                                 }
                                                                                  }
                                                                              ],
                                                               "requestBody":  {
                                                                                   "content":  {
                                                                                                   "application/json":  {
                                                                                                                            "schema":  {
                                                                                                                                           "$ref":  "#/components/schemas/UpdateStatusRequest"
                                                                                                                                       }
                                                                                                                        }
                                                                                               },
                                                                                   "required":  true
                                                                               },
                                                               "responses":  {
                                                                                 "200":  {
                                                                                             "description":  "OK",
                                                                                             "content":  {
                                                                                                             "*/*":  {
                                                                                                                         "schema":  {
                                                                                                                                        "$ref":  "#/components/schemas/ApiResponseUserResponse"
                                                                                                                                    }
                                                                                                                     }
                                                                                                         }
                                                                                         }
                                                                             },
                                                               "security":  [
                                                                                {
                                                                                    "bearerAuth":  [

                                                                                                   ]
                                                                                }
                                                                            ]
                                                           }
                                             },
                  "/api/users/{id}/role":  {
                                               "patch":  {
                                                             "tags":  [
                                                                          "User Management"
                                                                      ],
                                                             "summary":  "Update user role",
                                                             "operationId":  "updateRole",
                                                             "parameters":  [
                                                                                {
                                                                                    "name":  "id",
                                                                                    "in":  "path",
                                                                                    "required":  true,
                                                                                    "schema":  {
                                                                                                   "type":  "integer",
                                                                                                   "format":  "int64"
                                                                                               }
                                                                                }
                                                                            ],
                                                             "requestBody":  {
                                                                                 "content":  {
                                                                                                 "application/json":  {
                                                                                                                          "schema":  {
                                                                                                                                         "$ref":  "#/components/schemas/UpdateRoleRequest"
                                                                                                                                     }
                                                                                                                      }
                                                                                             },
                                                                                 "required":  true
                                                                             },
                                                             "responses":  {
                                                                               "200":  {
                                                                                           "description":  "OK",
                                                                                           "content":  {
                                                                                                           "*/*":  {
                                                                                                                       "schema":  {
                                                                                                                                      "$ref":  "#/components/schemas/ApiResponseUserResponse"
                                                                                                                                  }
                                                                                                                   }
                                                                                                       }
                                                                                       }
                                                                           },
                                                             "security":  [
                                                                              {
                                                                                  "bearerAuth":  [

                                                                                                 ]
                                                                              }
                                                                          ]
                                                         }
                                           },
                  "/api/users":  {
                                     "get":  {
                                                 "tags":  [
                                                              "User Management"
                                                          ],
                                                 "summary":  "Get all users",
                                                 "operationId":  "getAllUsers",
                                                 "responses":  {
                                                                   "200":  {
                                                                               "description":  "OK",
                                                                               "content":  {
                                                                                               "*/*":  {
                                                                                                           "schema":  {
                                                                                                                          "$ref":  "#/components/schemas/ApiResponseListUserResponse"
                                                                                                                      }
                                                                                                       }
                                                                                           }
                                                                           }
                                                               },
                                                 "security":  [
                                                                  {
                                                                      "bearerAuth":  [

                                                                                     ]
                                                                  }
                                                              ]
                                             }
                                 },
                  "/api/users/{id}":  {
                                          "get":  {
                                                      "tags":  [
                                                                   "User Management"
                                                               ],
                                                      "summary":  "Get user by ID",
                                                      "operationId":  "getUserById",
                                                      "parameters":  [
                                                                         {
                                                                             "name":  "id",
                                                                             "in":  "path",
                                                                             "required":  true,
                                                                             "schema":  {
                                                                                            "type":  "integer",
                                                                                            "format":  "int64"
                                                                                        }
                                                                         }
                                                                     ],
                                                      "responses":  {
                                                                        "200":  {
                                                                                    "description":  "OK",
                                                                                    "content":  {
                                                                                                    "*/*":  {
                                                                                                                "schema":  {
                                                                                                                               "$ref":  "#/components/schemas/ApiResponseUserResponse"
                                                                                                                           }
                                                                                                            }
                                                                                                }
                                                                                }
                                                                    },
                                                      "security":  [
                                                                       {
                                                                           "bearerAuth":  [

                                                                                          ]
                                                                       }
                                                                   ]
                                                  }
                                      },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "UserResponse":  {
                                                            "type":  "object",
                                                            "properties":  {
                                                                               "id":  {
                                                                                          "type":  "integer",
                                                                                          "format":  "int64"
                                                                                      },
                                                                               "name":  {
                                                                                            "type":  "string"
                                                                                        },
                                                                               "email":  {
                                                                                             "type":  "string"
                                                                                         },
                                                                               "role":  {
                                                                                            "type":  "string",
                                                                                            "enum":  [
                                                                                                         "VIEWER",
                                                                                                         "ANALYST",
                                                                                                         "ADMIN"
                                                                                                     ]
                                                                                        },
                                                                               "status":  {
                                                                                              "type":  "string",
                                                                                              "enum":  [
                                                                                                           "ACTIVE",
                                                                                                           "INACTIVE"
                                                                                                       ]
                                                                                          },
                                                                               "createdAt":  {
                                                                                                 "type":  "string",
                                                                                                 "format":  "date-time"
                                                                                             }
                                                                           }
                                                        },
...
                                       "UpdateStatusRequest":  {
                                                                   "required":  [
                                                                                    "status"
                                                                                ],
                                                                   "type":  "object",
                                                                   "properties":  {
                                                                                      "status":  {
                                                                                                     "type":  "string",
                                                                                                     "enum":  [
                                                                                                                  "ACTIVE",
                                                                                                                  "INACTIVE"
                                                                                                              ]
                                                                                                 }
                                                                                  }
                                                               },
                                       "ApiResponseUserResponse":  {
                                                                       "type":  "object",
                                                                       "properties":  {
                                                                                          "success":  {
                                                                                                          "type":  "boolean"
                                                                                                      },
                                                                                          "message":  {
                                                                                                          "type":  "string"
                                                                                                      },
                                                                                          "data":  {
                                                                                                       "$ref":  "#/components/schemas/UserResponse"
                                                                                                   },
                                                                                          "timestamp":  {
                                                                                                            "type":  "string",
                                                                                                            "format":  "date-time"
                                                                                                        }
                                                                                      }
                                                                   },
                                       "UpdateRoleRequest":  {
                                                                 "required":  [
                                                                                  "role"
                                                                              ],
                                                                 "type":  "object",
                                                                 "properties":  {
                                                                                    "role":  {
                                                                                                 "type":  "string",
                                                                                                 "enum":  [
                                                                                                              "VIEWER",
                                                                                                              "ANALYST",
                                                                                                              "ADMIN"
                                                                                                          ]
                                                                                             }
                                                                                }
                                                             },
                                       "ApiResponseListUserResponse":  {
                                                                           "type":  "object",
                                                                           "properties":  {
                                                                                              "success":  {
                                                                                                              "type":  "boolean"
                                                                                                          },
                                                                                              "message":  {
                                                                                                              "type":  "string"
                                                                                                          },
                                                                                              "data":  {
                                                                                                           "type":  "array",
                                                                                                           "items":  {
                                                                                                                         "$ref":  "#/components/schemas/UserResponse"
                                                                                                                     }
                                                                                                       },
                                                                                              "timestamp":  {
                                                                                                                "type":  "string",
                                                                                                                "format":  "date-time"
                                                                                                            }
                                                                                          }
                                                                       },
...
                                   },
                       "securitySchemes":  {
                                               "bearerAuth":  {
                                                                  "type":  "http",
                                                                  "description":  "JWT Bearer Token. Format: Bearer \u003ctoken\u003e",
                                                                  "in":  "header",
                                                                  "scheme":  "bearer",
                                                                  "bearerFormat":  "JWT"
                                                              }
                                           }
                   }
}