# src/main/java/com/financeapp/dto/request/UpdateRoleRequest.java
package com.financeapp.dto.request;

import com.financeapp.enums.Role;
import jakarta.validation.constraints.NotNull;
import lombok.Data;

@Data
public class UpdateRoleRequest {

    @NotNull(message = "Role is required")
    private Role role;
}