# src/main/java/com/financeapp/dto/request/UpdateStatusRequest.java
package com.financeapp.dto.request;

import com.financeapp.enums.UserStatus;
import jakarta.validation.constraints.NotNull;
import lombok.Data;

@Data
public class UpdateStatusRequest {

    @NotNull(message = "Status is required")
    private UserStatus status;
}